Commits:

1. LESS.js v1.3.3 to v2.5.3 (see below some benchmark testing)
 * I have add'd color customization in my plugin and some color functions are not working correctly, a new version would solve this issue's.
1. Remove self closing XHTML and make it HTML5 valid (nitpicking here :+1:)
1. Do not load (or generate) empty dynamic LESS/CSS file:
 * I noticed if I supply the webassets Bundle an empty array that it would not parse the block but it does remove the block from parsing in the final HTML and therefor not loading, parsing or generating any empty files.
 
----------
**Benchmark of LESS v1.3.3 vs v2.5.3**
The file of 2.5.3 is bigger but the processing is faster and therefor most browsers had there DOM loaded first with 2.5.3

**Old library (v1.3.3)**

1. Chrome
Total page load 7.5s (6.87s)
Empty less has finished. css generated in 61ms
TouchUI less has finished. css generated in 780ms

1. Mac chrome (with SauceConnect Tunnel)
Total page load 12s (8.3s)
Empty less has finished. css generated in 279ms
TouchUI less has finished. css generated in 1785ms

1. Firefox
Total page load 10s
Empty less has finished. css generated in 488ms
TouchUI less has finished. css generated in 858ms

1. Mac Firefox (with SauceConnect Tunnel)
Total page load 11s
Empty less has finished. css generated in 700ms
TouchUI less has finished. css generated in 1700ms

1. IE Edge
Total page load 11s (6.3s Tunnel)
Empty less has finished. css generated in 36ms
TouchUI less has finished. css generated in 832ms

1. Safari (with SauceConnect Tunnel)
Total page load 12s (7s)
Empty less has finished. css generated in 350ms
TouchUI less has finished. css generated in 3900ms

**New library (v2.5.3)**

1. Chrome
Total page load 7.5s (6.23s)
Empty less has finished. css generated in 92ms
TouchUI less has finished. css generated in 715ms

1. Mac chrome (with SauceConnect Tunnel)
Total page load 13s (8s)
Empty less has finished. css generated in 403ms
TouchUI less has finished. css generated in 1815ms

1. Firefox
Total page load 10s
Empty less has finished. css generated in 761ms
TouchUI less has finished. css generated in 890ms

1. Mac Firefox (with SauceConnect Tunnel)
Total page load 10s
Empty less has finished. css generated in 1100ms
TouchUI less has finished. css generated in 1600ms

1. IE Edge
Total page load 11s (6.7s)
Empty less has finished. css generated in 86ms
TouchUI less has finished. css generated in 912ms

1. Safari (with SauceConnect Tunnel)
Total page load 12s (6.5s)
Empty less has finished. css generated in 350ms
TouchUI less has finished. css generated in 4000ms