@foosel 2015-12-27
Thanks for this! Currently on vacation, will review and merge after new years, hope you understand :)